requirements.txt
django-allauth
git+https://github.com/InQuant/django-forms-bs